To disable service even when it is in unknown state
A service in unknown state is currently not reacting to disable event.
It prevents the service group from being disabled->go-standby. When a
service is in unknown state, the swact is blocked on the active
controller.

This fix allows the attempt to disable a service in unknown state.
The swact will be determined by the result of the disable action, i.e,
failing to disable a critical service will fail the swact. The swact
will continue if disabling a major or minor service fails, with alarm
or warning.

Closes-bug: 1831273

Change-Id: I30010c66c60fcf0eec8399a35e159755a1d3d318
Signed-off-by: Bin Qian <bin.qian@windriver.com>