Enhance timer system to avoid double deregister
The bug reported was because the dbmon service audit timer was
overwritten accidentally, therefore no audit was performed so the
dbmon service was not actually being audit.

Major change is to enhance timer system to use global unique timer
id (not reused) to ensure timer is not double deregistered by 2
different mechanisms (disarm/deregister).
Change the timer id to 64 bit integer to ensure id never overflow.

Above change eliminates the double deregistering a timer issue which
could accidentally deregister a new timer that reuses the same id.

Also some cleaning to get rid of cases that could double deregister
timer (although it is no longer harmful as above mentioned change is
in place)

Change-Id: I2603870d2eb2749d78456e406095ae543353963f
Closes-Bug: 1837724
Signed-off-by: Bin Qian <bin.qian@windriver.com>