Remove not needed shebangs in sm_client
The linters in the Opensuse build service are failing because sm_client has
unneeded python shebangs in the code. This is because a python source code
file that is not intended to be executed shouldn't include this shebang.

Also, the linter fails as `/usr/bin/env python` is used causing that the
dependency discovery tool fails. It is safe to use `/usr/bin/python` as
currently we don't provide any other python version.

Story: 2006508
Task: 36647

Change-Id: If3f83b9562414c3392515828a3c716a5bc23015d
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>