Fix LSB headers in sm
The opensuse build system reported two linter issues regarding
the LSB scripts in sm. The issues are:

  - For `sm`: Has `Should-Start` but no `Should-Stop`.
  - For `sm-shutdown`: LSB header not found.

To fix this issues the `Should-Stop` line was added in `sm` and
the LSB header was added in `sm.shutdown` script.

In `sm.shutdown` the `Default-Start` and `Default-Stop` were set
as the same as `sm`. `sm.shutdown` does nothing on the start stage
so this change won't affect any functionality.

Story: 2006508
Task: 36648

Change-Id: I4fac67a0a1c1abd82e47a3293aeae3036ee9722b
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>