Bug1845393 remove interface recovering state
In the case of a switch recycle, the connected nic will go down and up
but the communication will restore after the switch is up and running.
This could take a few seconds (much longer than anticipated).

This holds off the i/f state update to the peer.

Also remove the batching interface failover state change. This is already
handled in the failover fsm fail_pending state.

Change-Id: Ia810927dbbc4b3821f7915e6a42bceeac43d9e46
Closes-Bug: 1845393
Signed-off-by: Bin Qian <bin.qian@windriver.com>