Reduce SM timeouts for sysinv-conductor
SM runs ocf script to disable sysinv-conductor. Currently,
the timeout for sysinv-conductor ocf script to forcibly
terminate sysinv-conductor is 55s and the timeout for SM
to kill sysinv-conductor ocf script is 60s. In the case
that host-swact happens during application and k8s upgrade
operations, it will take a long time to complete swact
as these operations may spawn long-running processes or
greenthreads which prevent the sysinv-conductor from
shutting down before reaching the timeout to forcibly shutdown.
In a controller swact scenario, this results in the system
having almost no services running for almost a minute,
while SM is waiting for sysinv-conductor to shut down.

This commit updates the timeout for sysinv-conductor ocf
script to the default 15s and reduces the timeout for SM
to kill the script is 20s.

Tests conducted:
  - platform sanity
  - perform host-swact during stx-openstack upload/apply/remove
  - perform host-swact during kube upgrade operations
  - perform host-swact during platform-integ-apps apply
  - verified sysinv-conductor can be forcely shutdown after 15s
    when long-running processes are running

Change-Id: I337154a140f6cec3d6ab953003bf355b4396249e
Story: 2006781
Task: 38480
Signed-off-by: Angie Wang <angie.wang@windriver.com>