Use newer flake8 to run on ubuntu-focal Zuul machines
flake8 2.5.5  fails on ubuntu-focal zuul machines running python3.8
with the following error:
AttributeError: 'FlakesChecker' object has no attribute 'CONSTANT'

The update removes the version constraint to use newer flake8.

The linters can be run in python3.
Pylint cannot be run in python3 because mysql-python is not
compatable, so a new zuul job for pylint is now added.

The flake8 errors that the newer version raises are all suppressed,
and some of them should be addressed by someone with familiarity in
this repo.

Change-Id: Ida6447728d4175173c02130cb04a6013e4f966f9
Partial-Bug: 1895054
Signed-off-by: albailey <Al.Bailey@windriver.com>