Skip verifying h/w info for Not-In-Use interface
When a domain interface state is changed, hardware information is
verified to ensure the interface is OK to enter into the new state.

However, when an interface is entering into Not-In-Use state, it should
be always OK no matter what the h/w interface state is. Especially when
the interface is back on lo, in which case getting hardware information
will fail. This prevents moving interface to Not-In-Use state.

This change skip verifying h/w state if state of an interface is changed
to Not-In-Use.

This fix will also skip checking h/w information for lo interface and
always returns enabled = true.

Closes-Bug: 1898629

Signed-off-by: Bin Qian <bin.qian@windriver.com>
Change-Id: I709708bce622f52bf84fc3fec749f204cfeee533