Fix zuul jobs broken due to pip upversion
The install_command for docs, newnote and api-ref
needed to be overridden to not use upper constraints.

The bandit requirement needed to be made python3 only.

The bandit scan was failing, so it is now updated to
allow individual bandit failures to be suppressed in tox.ini

Need to include a py file change in order for bandit to be
triggered by zuul.

Partial-Bug: #1907678
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: Ic73d0ea590ab1b7857f7275fa9c71828b0d343ee