Add support for dcmanager-audit-worker service
We're moving the bulk of the dcmanager subcloud audits to separate
worker processes, so we need to add a service for the main worker
processes (which will then spawn additional workers).

In order to ensure that audits can be processed as soon as
dcmanager-audit starts up, we make enabling it dependent on
dcmanager-audit-worker being already running.

Story: 2007267
Task: 41870
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/769216
Change-Id: I162c00a3e8dba07f1912171e9371c29e5fd9a689