Python 3: Enable sm-api to work after unlock
The "min()" function works for None arguments
in Python 2 but in Python 3 raises an error.
We add a check to ensure that the function is not
called with the None argument in Python 3.

Story: 2006796
Task: 42388
Depends-On: I29a81755f732b55f67321748604b2e5d951935c9
Change-Id: I2199e0b87d342cec305edbb10483c6c038ad591c
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>
(cherry picked from commit f2186d04c0bed873a9f601c952c8b91442ca68c4)