sm-common: a temporary fix to build mtce prior to sm-common
sm-db build-depends on sm-common-libs which run-depends on mtce, if
sm-db builts prior to mtce, a failure is:

"sm-common-libs : Depends: mtce-pmon but it is not installable"

This is a temporary fix to make the build order as mtce, sm-common,
sm-db to bypass it.

Once it has been resolved on build system, the fix can be reverted.

Story: 2009101
Task: 44976

Test Plan:
Pass: build-pkgs -a without failure of sm-db.

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: I76c16dc87143db9b8a0484d74071afb6e8740745