Update SM lsb script for quick start
pidof command returns subprocess id when SM main process terminates.
This result a false postive that SM is already running so the start
action is skipped.

Make changes to the SM lsb script to distingrish if a subprocess ID
is returned, and attempt to kill it to speed up recovery of SM.

Revert the change to extend startuptime to 15 seconds back to 5.

Test Cases:
    kill SM process, observe SM process starts immediately after the
    subprocess is killed. SM is recovered within 2 seconds.
    (calculated by last and first logging of SM)

Change-Id: Ida834e7dd31a493ee6193b4d8ee73ebd97513de2
Closes-Bug: 1998349
Signed-off-by: Bin Qian <bin.qian@windriver.com>