Increase retries and timeouts on "audit-enabled" of mgr-restful-plugin in SM database
When the system is unstable, using a lot of CPU, it takes
more time for the communication between the components to happen.

So it's necessary to increase the maximum of retries and timeouts
in the "audit-enable" of the mgr-restful-plugin to prevent errors
from happening.

Test Plan:
PASS: mgr-restful-plugin restarted by SM (AIO-SX)

Closes-bug: 2000080

Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>
Change-Id: I0f8462fef20196a3bb913fa7d374a86a2c6565f1