Remove guest-agent related queries from sm database
The guest-agent service it is currently being activated
in setups where stx-openstack is applied but it's not
being used since we went to containerized openstack.
Since this service is no longer being used we are currently
removing the service and all related queries that are on
the create_sm_db file.

Test Plan:
PASS: Perform a fresh install on a duplex environment and
check that no error log related to guest-agent is appearing
in the sm log file and no 400.02 alarm was raised on
fm alarm-list.

Closes-Bug: 2003117

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I145bd8a45c12319facc4d1eff90b785a33a1d2c0