Remove disable dependence between ceph-manager and sysinv-conductor
When the system is unstable, using a lot of CPU, it takes
more time for the communication between the components to happen,
such as the communication between mgr-restful-plugin and ceph-mgr.
This communication failure may result on a failed audit by SM,
which then restarts the mgr-restful-plugin.

Change [1] resolved this issue by increasing the timeout and retries
of mgr-restful-plugin in SM database.

However, there is a disable dependence chain between mgr-restful,
ceph-manager, and sysinv-conductor which results on sysinv-conductor
being restarted if mgr-restful-plugin or ceph-manager is also disabled
by SM. This can impact platform-integ-apps apply or any other action
being executed by sysinv-conductor.

The ceph manager -> sysinv-conductor dependence is not necessary
anymore after the changes [2] and [3] were merged, thus this change
removes this dependence.

TEST PLAN:
PASS: AIO-SX: bootstrap, unlock and apply platform-integ-apps
PASS: Force ceph-manager to be restarted by SM, and verify that
      sysinv-conductor keeps running

Related-Bug: 2000080

[1] https://review.opendev.org/c/starlingx/ha/+/868118
[2] https://review.opendev.org/c/starlingx/utilities/+/856320
[3] https://review.opendev.org/c/starlingx/utilities/+/860570

Signed-off-by: Alyson Deives Pereira <alyson.deivespereira@windriver.com>
Change-Id: I949ccebd509b8099870b3dfda252a60b6b423715