Adjusting SM ceph-mon service parameters
Adjusting the ceph-mon service to call the scripts with a new
parameter. The old one was just 'mon' and the new one must be
'mon.controller'.

This change is needed for the AIO-DX to support Ceph with 3 ceph
monitors running. It prevents the ceph-init-wrapper script from
managing the other fixed monitors.

A new service called storage-networking is added to improve the Ceph
resiliency for AIO-DX when there is no networking communication
between the controllers, avoiding a split-brain condition.

Test-Plan:
  PASS: Fresh install and check if only the floating monitor is being
diabled/enabled on the controllers after a swact. The other monitors
should not be affected by a swact.

Story: 2011122
Task: 50128

Depends-On: https://review.opendev.org/c/starlingx/integ/+/914913

Change-Id: I5731e2ac3cf726eed765645b95c392ea55b1c94f
Signed-off-by: Hediberto C Silva <hediberto.cavalcantedasilva@windriver.com>