Fix swact back after legacy upgrade
This commit is to change the swact precheck logic. The new
logic will unblock the swact precheck for host state sync state
if the USM endpoint is not present. This change is needed for
the release that doesn't have USM endpoint present.

Test Plan:

PASS: swact back to controller 0 after legacy upgrade to 24.09
PaSS: swact between controllers in 24.09

Task: 49826
Story: 2010676
Change-Id: I6824f00589057f4d48c8df04425431cb22361e8e
Signed-off-by: junfeng-li <junfeng.li@windriver.com>