Fix ipsec-config service dependency
This commit sets a dependency between ipsec-config service and
management-ipv4/ipv6 services. The disable action may be performed on
ipsec-config service after management-ipv4/ipv6 dependent services are
on disabled state.

This fix is needed due to the verification step present on monitor
function for ipsec-config service. This function checks if the
floating IP is present on system ip tables and swanctl configuration
and verifies if the conditions are satisfied for active and standby
controllers.

It is expected that floating IP is present on the active controller
system where ipsec-config is on enabled-active state and not present
on standby controller system where ipsec-config is on disabled state.
The floating IP is added and removed by management-ipv4/ipv6 service
per their start and stop actions. In the previous service dependency
configuration, this would cause an error during ipsec-config audit-
disabled action and 400.001 alarm was present on system.

Therefore, this commit fixes this service dependency relation between
ipsec-config and management-ip services.

Test Plan:
PASS: Full build, system install, bootstrap and unlock of a DX system
      with unlocked enabled available state. No 400.001 alarms present
      on system.
PASS: On a DX system with unlocked enabled available state, perform a
      host-swact on controller-0. Observe that ipsec-config service
      changes its state on controllers, from disabled to enabled-active
      on active controller and from enabled-active to disabled on
      standby controller. No errors are reported on daemon-ocf.log
      related to ipsec-config or management-ip services. No 400.001
      alarms present on system.
PASS: On a DX system with unlocked enabled available state, perform a
      host-lock and host-unlock on controller-1. Observe that system
      boots with ipsec-config service on disabled state. No errors are
      reported on daemon-ocf.log related to ipsec-config or
      management-ip services. No 400.001 alarms present on system.

Story: 2010940
Task: 50196

Change-Id: Idd2f487b4589e1f66d79ea5c4f13c36e67c302be
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>