Remove /var/run/.sm_reset_peer flag file after its 30 second timeout
When the SM on one controller detects a process stall on its peer
via the maintenance heartbeat cluster, it attempts recovery by
signaling maintenance to reset the peer. The signal to do so is
by creating the /var/run/.sm_reset_peer flag file.

Normally, if the peer's BMC is provisioned, the mtcClient accepts the
reset request, performs the reset, and then removes the
/var/run/.sm_reset_peer flag file.

Once SM has created this flag file it waits for up to 30 seconds for
mtcClient to remove it which is the signal that the reset is performed.

However, if the peer's BMC is not provisioned, the mtcClient cannot
perform the reset. In this case SM times out and moves on but leaves
the file present.

While the flag file (signal) remains present then the mtcClient is
poised to reset the peer controller when the peer's BMC is provisioned.

To avoid that future unexpected reset this update has SM remove the
/var/run/.sm_reset_peer flag file if it times out.

Test Plan: This update only affects the timeout handling case

PASS: Verify flag file is removed on timeout case.
PASS: Verify on timeout case that the mtcClient sees the file as gone.

PASS: Verify flag file remove operation is handled gracefully if
      the file is already gone. Handle race condition case where
      mtcClient removes it at timeout but before SM tries to remove.

Closes-Bug: 2076454
Change-Id: I694f1107e86fc90c7f007918fb59e754482717fe
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>