SM service logs are being duplicated into postgres.log
The logging configuration update for the SM service logs
is invalid and conflicts with the postgres logging
facility (local0), causing the SM logs to appear in
both sm-service.log and postgres.log. Reverting to
facility (local3) for SM service logs to fix the issue.

Test Plan: Verified the changes on simplex and duplex configurations.

PASS: AIO-Simplex - Verified service logs are written to sm-service.log
                    Verfied that service logs are not seen in sm.log
                    and postgres.log
PASS: AIO-Duplex -  Verified service logs are written to sm-service.log
                    Verified that service logs are not in sm.log and
                    postgres.log
PASS: AIO-Duplex -  Verified that service logs are written to
                    sm-service.log file after a controlled swact and
                    not seen in sm.log and postgres.log
Change-Id: I03b3b033e38801711d1e6ab6831913c5e5b067a4
Signed-off-by: Sandhya Kalisetty <sandhya.kalisetty@windriver.com>