Support YAML oslo policy file in sm-api
The usage of JSON oslo policy files was deprecated in Victoria [1]. This
change converts the current policy file of sm-api to YAML and adds the
support to load YAML files to the internal implementation of the policy
parser.

The policy file was converted to YAML, using the tool provided by oslo
('oslopolicy-convert-json-to-yaml').

[1] https://docs.openstack.org/releasenotes/oslo.policy/victoria.html

Test Plan:
PASS: Deploy AIO-SX (over master branch).
      Check that the YAML files are present.
PASS: Check the YAML rules using the oslo tool 'oslopolicy-checker'.

Story: 2011360
Task: 53167

Change-Id: I0c9d129a1075f98ce001e47d75636e069037e6bc
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>